change: move Navigation out of Layout
