+++ src/components/Navigation/index.js
@@ -1,0 +1,2 @@
+import Navigation from "./Navigation";
+export default Navigation;
